cleanup sbt

+++ code/service/src/main/scala/com/namely/chiefofstate/Util.scala
@@ -28,7 +28,7 @@ object Util {
   def toCosMetaData(metaData: MetaData): CosMetaData = {
     CosMetaData(
       entityId = metaData.entityId,
-      revisionNumber = metaData.revisionNumber,
+      revisionNumber = metaData.revisionNumber.toInt,
       revisionDate = metaData.revisionDate,
       data = metaData.data
     )
